优化了import

+++ src/signal/signal_emitter.py
@@ -5,6 +5,6 @@ from PySide6.QtWidgets import QListWidgetItem
 class SignalEmitter(QObject):
     parse_finished: Signal = Signal(tuple)
     download_start: Signal = Signal(object, object, object)
-    progress_update: Signal = Signal(int)
+    progress_update: Signal = Signal(float)
     download_finished: Signal = Signal(QListWidgetItem)
     check_update: Signal = Signal(str, str, bool, str)
